requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
dateutils==0.6.6
idna==2.6
lxml==3.8.0
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.13
urllib3==1.22
